requirements.txt
# python 기본패키지
python-dotenv>=1.1.1

# FastAPI & 웹서버
fastapi>=0.116.0
uvicorn>=0.35.0

# google Gemini
google-genai>=1.25.0